+++ BOOT/product/src/main/java/com/coopang/product/domain/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.coopang.product.domain.repository;
 
-import com.coopang.product.infrastructure.entity.CategoryEntity;
+import com.coopang.product.domain.entity.CategoryEntity;
 import java.util.Optional;
 import java.util.UUID;
