+++ BOOT/product/src/main/java/com/coopang/product/infrastructure/repository/CategoryJpaRepository.java
@@ -1,6 +1,6 @@
 package com.coopang.product.infrastructure.repository;
 
-import com.coopang.product.infrastructure.entity.CategoryEntity;
+import com.coopang.product.domain.entity.CategoryEntity;
 import com.coopang.product.domain.repository.CategoryRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
